Module Ezcurl_core

Core signatures and implementation

module Config : sig ... end

Configuration for the client.

type t = private {

curl : Curl.t;

}

A client, i.e. a cURL instance. The wrapping record has been present since 0.3

val make : ?set_opts:(Curl.t -> unit) -> ?cookiejar_file:string -> ?enable_session_cookies:bool -> unit -> t

Create a new client.

parameter set_opts
called before returning the client, to set options

parameter cookiejar_file
if provided, tell curl to use the given file path to store/load cookies (since 0.3)

parameter enable_session_cookies
if provided, enable cookie handling in curl so it store/load cookies (since 0.3)

val delete : t -> unit

Delete the client. It cannot be used anymore.

val with_client : ?set_opts:(Curl.t -> unit) -> (t -> 'a) -> 'a

Make a temporary client, call the function with it, then cleanup.

val set_no_signal : bool -> unit

Set no_signal default value for each new client instance. Default is true. See CURLOPT_NOSIGNAL.

module Cookies : sig ... end

Cookie handling.

type response_info = {

ri_response_time : float;
(*
Total time (in seconds) for the request/response pair. See Curl.get_totaltime.
*)

ri_redirect_count : int;
(*
Number of redirects cURL followed. See Curl.get_redirectcount.
*)

}

Metadata about a response from the server.

val pp_response_info : Stdlib.Format.formatter -> response_info -> unit

val string_of_response_info : response_info -> string

type 'body response = {

code : int;
(*
Response code. See https://developer.mozilla.org/en-US/docs/Web/HTTP/Status
*)

headers : (string * string) list;
(*
Response headers
*)

body : 'body;
(*
Response body, or ""
*)

info : response_info;
(*
Information about the response
*)

}

Response for a given request.

val pp_response_with : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a response -> unit

val pp_response : Stdlib.Format.formatter -> string response -> unit

val string_of_response : string response -> string

type meth =

| GET

| POST of Curl.curlHTTPPost list

| PUT

| DELETE

| HEAD

| CONNECT

| OPTIONS

| TRACE

| PATCH

| HTTP_CUSTOM of string
(*
Custom HTTP method (e.g., for WebDAV)
*)

The HTTP method to use

val pp_meth : Stdlib.Format.formatter -> meth -> unit

val string_of_meth : meth -> string

module type IO = sig ... end

module type S = sig ... end

module Make (IO : IO) : S with type 'a io = 'a IO.t