Module Ezcurl_core.Cookies

Cookie handling.

val flush_cookiejar : t -> unit

If cookiejar_file was provided in make, this flushes the current set of cookies to the provided file.

val reload_cookiejar : t -> unit

If cookiejar_file was provided in make, this reloads cookies from the provided file.

val get_cookies : t -> string list

Get cookie list (in netscape format)

val set_cookies : t -> string list -> unit

Set cookie list (in netscape format)

val transfer : t -> t -> unit

transfer c1 c2 copies cookies in c1 into c2